+++ docusaurus.config.ts
@@ -77,7 +77,7 @@ const config: Config = {
             from: "/ios/upgrade-guide/",
           },
           {
-            to: "/ios/5x/features/tracing/",
+            to: "/ios/5x/features/traces/",
             from: "/ios/features/performance-monitoring/",
           },
           {
@@ -117,7 +117,7 @@ const config: Config = {
             from: "/ios/features/push-notifications/",
           },
           {
-            to: "/ios/5x/features/tracing/",
+            to: "/ios/5x/features/traces/",
             from: "/ios/features/tracing/",
           },
           {
@@ -173,7 +173,7 @@ const config: Config = {
             from: "/ios/integration/update-embrace/",
           },
           {
-            to: "/android/features/tracing/",
+            to: "/android/features/traces/",
             from: "/android/features/performance-monitoring/",
           },
           {
@@ -181,8 +181,40 @@ const config: Config = {
             from: "/open-telemetry/error-logs-tracking",
           },
           {
-            to: "/features/performance-tracing",
+            to: "/features/traces",
             from: "/product/performance-tracing",
+          },
+          {
+            to: "/android/features/traces",
+            from: "/android/features/tracing",
+          },
+          {
+            to: "/flutter/features/traces",
+            from: "/flutter/features/tracing",
+          },
+          {
+            to: "/ios/5x/features/traces",
+            from: "/ios/5x/features/tracing",
+          },
+          {
+            to: "/ios/open-source/traces",
+            from: "/ios/open-source/tracing",
+          },
+          {
+            to: "/react-native/4x/features/traces",
+            from: "/react-native/4x/features/tracing",
+          },
+          {
+            to: "/react-native/features/traces",
+            from: "/react-native/features/tracing",
+          },
+          {
+            to: "/unity/features/traces",
+            from: "/unity/features/performance-tracing",
+          },
+          {
+            to: "/features/traces",
+            from: "/features/performance-tracing",
           },
         ],
       },